// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/abwab-door-relation-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { AbwabRelationDirection } from '../models/abwab-relation-direction';
import { AbwabRelationType } from '../models/abwab-relation-type';
export interface AbwabDoorRelationDto {
  direction: AbwabRelationDirection | null;
  id: number;
  otherDoorId: number;
  otherDoorName: string;
  type: AbwabRelationType;
}